Changed the structure of the project. The module in charge of the simulation is now contained in doubleslit/cranknicolson and the GUI of the application is found in doubleslit. In cn2d.py now psi0 and V are passed as 2d arrays instead of functions.
