+++ doubleslit/cranknicolson/V2D.py
@@ -1,0 +1,7 @@
+import numpy as np
+import matplotlib.pyplot as plt
+
+V = np.loadtxt("V2d.dat")
+
+plt.imshow(V)
+plt.show()
