+++ doubleslit/doubleslit.kv
@@ -1,0 +1,28 @@
+#:kivy 1.0.9
+#: import main doubleslit
+
+
+<DoubleSlitScreen>:
+    p_rectangle: p_rect
+
+    orientation: "vertical"
+    padding: 10
+    spacing: 10
+
+    Widget:
+        size_hint: (1, 0.5)
+        id: p_rect
+
+    BoxLayout:
+        orientation: "horizontal"
+        spacing: 50
+        padding: [50, 50, 50, 50]
+
+        Button:
+            text: "Compute"
+            on_press: root.compute()
+        Button:
+            text: "Play/Pause"
+            on_press: root.playpause()
+        Button:
+            text: "Stop"
